improve array enumeration in megaclisas-status
Prior to this change this script would fail if a controller has non
consecutive arrays because the script expects the array id to start at 0
and increment to the array count.

With this change we retrieve a list of all arrays per controller and
iterate through that way.

I believe this situation occurs when an array is deleted, in our
situation array 0 was removed leaving just 1 & 2

Fixes #26